Clean-up and separate checkpoints vs snapshots
